add jacobian computation
This algorithm is based on the fact that the jacobian for one body
is the jacobian from its parent body (with a transform) plus the
contribution of the joint which joins it to its parent.

Instead of multiplying the whole parent jacobian with the transform, we
exploit the (branch induced) sparsity of the jacobian and only multiply
the non-zero blocks.

Some refactoring, renaming is much probably needed.

The jacobian is in R^(6*nbbodies x nbdof) and is stored in Robot::J.

alternatively, I could have stored nbbobies jacobians in R^(6 x nbdof) in
each Body::J. I do not know if one choice is better thant the other. One
drawback of the current solution is that it requires us to define
Robot::NBBODIES.

The results have been hand-checked for the simple-arm model but not for
the simple-humanoid model.