implement rnea with a visitor
The implementation is rather contrived because we need to
differentiate between root bodies and others.

We could make it more simple by changing depth_first_traversal
so that visitors take two template parameters: Node and PrevNode
similarly to backward_traversal.

PrevNode would be the parent of Node. It would le us specialize
easily the case of a root node:

    template <Node, PrevNode> struct Visitor;
    template <Node> struct Visitor<Node, NP> {};

Yet this modification would not cover the child-less node special
case:

on can imagine that in a recursive algorithm, when visiting a node
some work is done to prepare the children visit. It would be great to
skip this work when the node has no child.