dllexport/import joint and body classes
Before this commit, only METAPOD_DLL{API,IMPORT,EXPORT} was defined.
However, there was several problems

* DLLEXPORT was used in place of DLLAPI

* Joint and Body classes were not exported

* we do not create any metapod DLL but instead simple_arm DLL and
  simple_humanoid DLL. So we need to pass the library name to macros
  which define DLL-exported classes.

This commit adds DLL API support for these two libs (and also others for
the models we generate for the benchmarks).

Support for metapod DLL API is kept for the common part, but I'm not sure it
is working nor wise. What would happen if an application links with both
simple_arm and simple_humanoid? Would duplicate (though identical) definitions
create problems?

This commit also moves the cmake code related simple_arm and simple_humanoid
to each model directory. Before, it was split between the project root
(for files listing) and the tests directory (for actual library
creation).

Moving this code was necessary in order to generate the libraries
config.hh (the directory acts as a namespace). I moved everything for
consistency. So the new CMakeLists.txt are there

    /include/metapod/models/simple-{arm,humanoid}/CMakeLists.txt

If having cmake code below /include is inappropriate, we could move
them in

    /simple-{arm,humanoid}/CMakeLists.txt

instead. In the future, I'd like to move the models out of the
metapod namespace. In such a case these last locations would make sense.
Note that cureently, there is also .cc files in
/include/metapod/models/simple-{arm,humanoid}.

Eventually, this commit adds an helper function to generate config.hh.
One cannot use the one provided by JRL-CMAKE as it can only be called from the
project root directory.