use embedfile in robotbuilder
to embed config.hh.in, then generate config.hh from it using regular
expressions. This is needed to allow the generation of models which are
complete and independent from [jrl-]cmake.

config.hh.in is copied from jrl-cmake, but modified to remove the
LIBRARY_VERSION macro.

This process (embedding template file and formating them with regex)
could be used for the other generated source files too. It would be easier
to maintain than the current stream-based solution. One could even imagine
letting the user pass the template.