remplace example models source with generated code
The new version of the code was generated by metapodfromurdf, from
the models descriptions in URDF files.

The simple arm model is unchanged.

The simple humanoid model has been slightly modified. While the
new version is equivalent to the former, the orientation of the
link frames and the joint axis are not always identical.

Has a consequence, the model, whose all joints were of type
REVOLUTE_AXIS_X, now has a mixture of REVOLUTE_AXIS_X (when the axis
is X) and REVOLUTE_AXIS_ANY (when the axis is Y or Z) joints.

This might have a small impact on performance.

If it is a problem, we can either add support for REVOLUTE_AXIS_{Y,Z}
or a feature to metapodfromurdf to enable it to rotate the model and
ensure all axis are aligned with X.

Also the reference data for the bcal test has been updated to match
the new frame definitions.

At last, note that the joint and body "label" member variables might
have changed also. They are apparently not used anyway.

[1] https://github.com/sbarthelemy/metapodfromurdf