WIP: add a fwd_kin algorithm
the plan for fwd_kin is to compute the forward kinematics and to remove
duplication and inefficiencies from jcalc, bcalc and rnea.

This first version adds a variation of the algorithm which does the computation
in two steps, where the first step does not consider the root joint.

It only works for models with a single tree.

Change-Id: I5c695c3059af4817c34a12176e68ea19d4d5167b